+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-# react-micro-frame
+# @micro-frame/react
 
 ## 0.0.1
 
 ### Patch Changes
 
-- 17fb963: Publish the initial release of react-micro-frame.
+- 17fb963: Publish the initial release of @micro-frame/react.